requirements.txt
pyyaml>=6.0
aiohttp>=3.9
lighter-sdk>=1.0.6
